requirements.txt
Flask~=1.1.2
SQLAlchemy~=1.4.5
WTForms~=2.3.3
Werkzeug~=1.0.1
python-dotenv~=0.17.0